Fix issue 131
http://code.google.com/p/sqljet/issues/detail?id=131

git-svn-id: http://svn.svnkit.com/repos/sqljet@936 06c6b4df-3a75-4b2a-9d74-3e57e8079485

+++ trunk/sqljet/src/org/tmatesoft/sqljet/core/internal/table/SqlJetBtreeDataTable.java
@@ -536,14 +536,20 @@ public class SqlJetBtreeDataTable extends SqlJetBtreeTable implements ISqlJetBtr
     }
 
     private boolean isRowIdExists(final long rowId) throws SqlJetException {
+        return isRowIdExists(rowId, true);
+    }
+
+    private boolean isRowIdExists(final long rowId, boolean keepPosition) throws SqlJetException {
         final long current = getRowId();
         if (rowId == current)
             return true;
         final boolean exists = goToRow(rowId);
-        if (current > 0) {
-            goToRow(current);
-        } else {
-            first();
+        if (keepPosition) {
+            if (current > 0) {
+                goToRow(current);
+            } else {
+                first();
+            }
         }
         return exists;
     }
@@ -571,7 +577,11 @@ public class SqlJetBtreeDataTable extends SqlJetBtreeTable implements ISqlJetBtr
         }
 
         if (Action.INSERT == action) {
-            existsRowId = isRowIdExists(rowId);
+            long oldRowId = getRowId();
+            existsRowId = isRowIdExists(rowId, false);
+            if (!existsRowId && oldRowId > 0) {
+                goToRow(oldRowId);
+            }
         } else if (Action.UPDATE == action) {
             if (currentRowId != rowId) {
                 existsRowId = isRowIdExists(rowId);
@@ -583,7 +593,9 @@ public class SqlJetBtreeDataTable extends SqlJetBtreeTable implements ISqlJetBtr
             case IGNORE:
                 return false;
             case REPLACE:
-                cursor.delete();
+                if (goToRow(rowId)) {
+                    cursor.delete();
+                }
                 break;
             default:
                 throw new SqlJetException(SqlJetErrorCode.CONSTRAINT, "Record with given ROWID already exists");
